refactor: standardize serverUrl naming and enhance key signing process by removing unused parameters and updating method visibility in DKLS and Schnorr classes
